Use short locale names 'se', 'dk', 'no' for default languages
