+++ apps/store/next.config.js
@@ -11,9 +11,6 @@ const nextConfig = {
   },
   productionBrowserSourceMaps: true,
   i18n,
-  publicRuntimeConfig: {
-    routingLocales: i18n.locales.filter((locale) => locale !== 'default'),
-  },
 }
 
 module.exports = nextConfig
