+++ apps/store/public/locales/dk/common.json
@@ -132,7 +132,7 @@
   "FOREVER_INTRO_READY_CTA": "Ja, giv mig et pristilbud",
   "FOREVER_LANDINGPAGE_INPUT_TEXT": "Indtast koden",
   "FOREVER_LANDINGPAGE_BTN_LABEL": "Næste",
-  "FOREVER_LANDINGPAGE_INFO_TEXT": "Dette er en eksklusiv hjemmeside, som kun er for inviterede personer. Hvis du ikke har en kode på nuværende tidspunkt, bliver du stillet videre til [vores almindelige hjemmeside](https:\/\/www.hedvig.com\/dk\/new-member).",
+  "FOREVER_LANDINGPAGE_INFO_TEXT": "Denne side er for inviterede personer. Hvis du ikke har en kode, kan du fortsætte til [indboforsikring](https:\/\/www.hedvig.com\/dk\/forsikringer\/indboforsikring), [husforsikring](https:\/\/www.hedvig.com\/dk\/forsikringer\/husforsikring) eller [studenterforsikring](https:\/\/www.hedvig.com\/dk\/forsikringer\/studenterforsikring) på vores almindelige hjemmeside.",
   "FOREVER_LANDINGPAGE_DESCRIPTION": "Tillykke! Du er blevet inviteret til Hedvig Forever. Brug min kode {CODE} til at sænke din månedlige pris. ",
   "FOREVER_DOWNLOAD_HEADING": "Velkommen {FIRST_NAME}!",
   "FOREVER_DOWNLOAD_PARAGRAPH": "Du er nu Hedvig-medlem. Download appen for at komme i gang, og husk at invitere dine venner!",
@@ -437,5 +437,19 @@
   "OFFER_PAGE_MISSING_MAIN_COVERAGE_ERROR": "Vælg mindst en hovedforsikring for at fortsætte",
   "OFFER_PAGE_SECTION_TITLE_MAIN": "Hovedforsikring",
   "OFFER_PAGE_SECTION_TITLE_ADDITIONAL": "Øvrige forsikringer",
-  "ONBOARDING_CROSS_SELL_HEADER": ""
+  "ONBOARDING_CROSS_SELL_HEADER": "",
+  "SIDEBAR_SWITCHING_NOTICE_HEADING": "",
+  "SIDEBAR_SWITCHING_NOTICE_BODY": "",
+  "SIDEBAR_SWITCHING_NOTICE_LINK": "",
+  "DOWNLOAD_PAGE_CAR_SWITCHER_SECTION_HEADING": "",
+  "DOWNLOAD_PAGE_CAR_SWITCHER_HEADING": "",
+  "DOWNLOAD_PAGE_CAR_SWITCHER_BODY": "",
+  "INITIATE_CAR_CANCELLATION_PAGE_TITLE": "",
+  "CAR_SWITCHER_TOO_SOON_HEADING": "",
+  "CAR_SWITCHER_TOO_SOON_BODY": "",
+  "CAR_SWITCHER_TOO_SOON_CONFIRM": "",
+  "LANDING_PAGE_UPDATE_CHOICES_SUBHEADING": "Du kan altid ændre dine valg senere.",
+  "OFFER_SWITCH_PROMOTION_PARAGRAPH_CAR_V2": "",
+  "SIDEBAR_SWITCHING_NOTICE_ADDENDUM": "",
+  "SIDEBAR_STARTDATE_CELL_LABEL_SWITCHER": ""
 }
