+++ apps/store/public/locales/no/common.json
@@ -132,7 +132,7 @@
   "FOREVER_INTRO_READY_CTA": "Ja, gi meg et prisforslag.",
   "FOREVER_LANDINGPAGE_INPUT_TEXT": "Fyll inn koden",
   "FOREVER_LANDINGPAGE_BTN_LABEL": "Neste",
-  "FOREVER_LANDINGPAGE_INFO_TEXT": "Dette er en eksklusiv nettside som er tiltenkt folk som er invitert. Hvis du for øyeblikket ikke har fått en invitasjon enda, så frykt ikke. Du kan fortsatt melde deg inn på vårt [vanlige nettsted.](https:\/\/www.hedvig.com\/no\/new-member)",
+  "FOREVER_LANDINGPAGE_INFO_TEXT": "Denne siden er tiltenkt folk som er invitert. Hvis du ikke har fått en invitasjon, kan du fortsette til [våre forsikringer](https:\/\/www.hedvig.com\/no\/forsikring) på vår vanlige nettside.",
   "FOREVER_LANDINGPAGE_DESCRIPTION": "Gratulerer! Du har blitt invitert til Hedvig Forever. Bruk min kode {CODE} for å senke din månedspris.",
   "FOREVER_DOWNLOAD_HEADING": "Velkommen {FIRST_NAME}!",
   "FOREVER_DOWNLOAD_PARAGRAPH": "Nå er du medlem av Hedvig. Last ned appen og kom igang. Ikke glem å invitere vennene dine!",
@@ -437,5 +437,19 @@
   "OFFER_PAGE_MISSING_MAIN_COVERAGE_ERROR": "Du må velge minst en hovedforsikring for å fortsette",
   "OFFER_PAGE_SECTION_TITLE_MAIN": "Hovedforsikring",
   "OFFER_PAGE_SECTION_TITLE_ADDITIONAL": "Tillegg",
-  "ONBOARDING_CROSS_SELL_HEADER": ""
+  "ONBOARDING_CROSS_SELL_HEADER": "",
+  "SIDEBAR_SWITCHING_NOTICE_HEADING": "",
+  "SIDEBAR_SWITCHING_NOTICE_BODY": "",
+  "SIDEBAR_SWITCHING_NOTICE_LINK": "",
+  "DOWNLOAD_PAGE_CAR_SWITCHER_SECTION_HEADING": "",
+  "DOWNLOAD_PAGE_CAR_SWITCHER_HEADING": "",
+  "DOWNLOAD_PAGE_CAR_SWITCHER_BODY": "",
+  "INITIATE_CAR_CANCELLATION_PAGE_TITLE": "",
+  "CAR_SWITCHER_TOO_SOON_HEADING": "",
+  "CAR_SWITCHER_TOO_SOON_BODY": "",
+  "CAR_SWITCHER_TOO_SOON_CONFIRM": "",
+  "LANDING_PAGE_UPDATE_CHOICES_SUBHEADING": "Du kan alltid oppdatere valgene dine senere.",
+  "OFFER_SWITCH_PROMOTION_PARAGRAPH_CAR_V2": "",
+  "SIDEBAR_SWITCHING_NOTICE_ADDENDUM": "",
+  "SIDEBAR_STARTDATE_CELL_LABEL_SWITCHER": ""
 }
