+++ apps/store/scripts/download-translations.sh
@@ -39,10 +39,10 @@ lokalise2 \
   --filter-langs 'en,da_DK,en_DK,sv_SE,en_SE,nb_NO,en_NO' \
   --language-mapping '[
     {"original_language_iso": "en","custom_language_iso": "default"},
-    {"original_language_iso": "da_DK","custom_language_iso": "da-dk"},
+    {"original_language_iso": "da_DK","custom_language_iso": "dk"},
     {"original_language_iso": "en_DK","custom_language_iso": "en-dk"},
-    {"original_language_iso": "sv_SE","custom_language_iso": "sv-se"},
+    {"original_language_iso": "sv_SE","custom_language_iso": "se"},
     {"original_language_iso": "en_SE","custom_language_iso": "en-se"},
-    {"original_language_iso": "nb_NO","custom_language_iso": "nb-no"},
+    {"original_language_iso": "nb_NO","custom_language_iso": "no"},
     {"original_language_iso": "en_NO","custom_language_iso": "en-no"}
   ]'
