+++ apps/store/src/lib/l10n/countries.ts
@@ -1,5 +1,4 @@
-import { toIsoLocale, locales } from '@/lib/l10n/locales'
-import { CountryCode, Language, Locale, CountryLabel, UiLocale, IsoLocale } from './types'
+import { CountryCode, Locale, CountryLabel, IsoLocale } from './types'
 
 export type CountryData = {
   id: CountryLabel
@@ -31,30 +30,4 @@ export const countries: Record<CountryLabel, CountryData> = {
     defaultLocale: Locale.DaDk,
     locales: [Locale.DaDk, Locale.EnDk],
   },
-}
-
-const localeCountries = Object.fromEntries(
-  Object.entries(countries).flatMap(([countryId, countryData]) =>
-    countryData.locales.map((locale) => [locale, countryId]),
-  ),
-) as Record<IsoLocale, CountryLabel>
-
-export const getCountryByLocale = (locale: UiLocale): CountryData => {
-  locale = toIsoLocale(locale)
-  const countryData = countries[localeCountries[toIsoLocale(locale)]]
-  if (!countryData) {
-    throw new Error(`Failed to find country by locale=${locale}`)
-  }
-  return countryData
-}
-
-export const getCountryLocale = (country: CountryLabel, language: Language): IsoLocale => {
-  const countryData = countries[country as CountryLabel]
-  if (!countryData) {
-    throw new Error(`Failed to find country id=${country}`)
-  }
-  return (
-    countryData.locales.find((locale) => locales[locale].language === language) ||
-    countryData.defaultLocale
-  )
 }
