+++ apps/store/src/lib/l10n/useCurrentCountry.ts
@@ -1,6 +1,6 @@
 import { useRouter } from 'next/router'
-import { getCountryByLocale } from './countries'
-import { getLocaleOrFallback } from './locales'
+import { getCountryByLocale } from '@/lib/l10n/countryUtils'
+import { getLocaleOrFallback } from '@/lib/l10n/localeUtils'
 
 export const useCurrentCountry = () => {
   const router = useRouter()
