+++ apps/store/src/lib/l10n/useCurrentLocale.ts
@@ -1,5 +1,5 @@
 import { useRouter } from 'next/router'
-import { getLocaleOrFallback } from './locales'
+import { getLocaleOrFallback } from '@/lib/l10n/localeUtils'
 
 export const useCurrentLocale = () => {
   const router = useRouter()
