+++ apps/store/src/routingLocales.js
@@ -1,3 +1,3 @@
 // Must not include 'default'
 /** @type {import('@/lib/l10n/types').RoutingLocale[]} */
-module.exports = ['sv-se', 'en-se', 'nb-no', 'en-no', 'da-dk', 'en-dk']
+module.exports = ['se', 'en-se', 'nb-no', 'en-no', 'da-dk', 'en-dk']
