--- apps/store/src/utils/isSupportedLocale.ts
@@ -1,9 +1,0 @@
-import { toRoutingLocale } from '@/lib/l10n/locales'
-import { Locale, UiLocale } from '@/lib/l10n/types'
-
-export const isSupportedLocale = (locale: unknown): locale is UiLocale => {
-  return (
-    typeof locale === 'string' &&
-    Object.values(Locale).some((x) => x === locale || toRoutingLocale(x) === locale)
-  )
-}
